# Simplify ListOfConstraintTypesPresent in Bridges